backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
SQLAlchemy
python-dotenv
requests
google-generativeai
